Refactored specular computation into Strategy classes.
No added functionality yet
Roughness model switchable prepared temporarily via MultiLayer::setRoughnessModel
