+++ Core/Computation/MultiLayerUtils.cpp
@@ -83,6 +83,26 @@ std::vector<ZLimits> MultiLayerUtils::ParticleRegions(const MultiLayer& multilay
     return layer_fill_limits.layerZLimits();
 }
 
+// TODO: test this function
+bool MultiLayerUtils::ContainsMagneticMaterial(const MultiLayer& sample)
+{
+    for (const Material* mat : sample.containedMaterials())
+        if (mat->isMagneticMaterial())
+            return true;
+    return false;
+}
+
+// TODO: test this function
+bool MultiLayerUtils::hasRoughness(const MultiLayer& sample)
+{
+    for(size_t i = 0; i < sample.numberOfLayers() - 1; i++ )
+    {
+        if( sample.layerInterface(i)->getRoughness() )
+            return true;
+    }
+    return false;
+}
+
 namespace
 {
 std::vector<double> BottomLayerCoordinates(const MultiLayer& multilayer)
