+++ Core/Multilayer/ISpecularStrategy.h
@@ -1,0 +1,34 @@
+
+
+#ifndef ISPECULARSTRATEGY_H
+#define ISPECULARSTRATEGY_H
+
+#include "ILayerRTCoefficients.h"
+//#include "ScalarRTCoefficients.h"
+//#include "MatrixRTCoefficients_v2.h"
+#include "Slice.h"
+
+//template<typename CoefficientsType>
+class ISpecularStrategy
+{
+public:
+
+//    virtual std::vector<ScalarRTCoefficients> eval(const std::vector<Slice>& slices,
+//                                                              kvector_t k) const = 0;
+
+//    virtual std::vector<ScalarRTCoefficients> eval(const std::vector<Slice>& slices,
+//                                                   const std::vector<complex_t>& kz) const = 0;
+
+//    virtual std::vector<MatrixRTCoefficients_v2> eval(const std::vector<Slice>& slices,
+//                                                              kvector_t k) const = 0;
+
+//    virtual std::vector<MatrixRTCoefficients_v2> eval(const std::vector<Slice>& slices,
+//                                                   const std::vector<complex_t>& kz) const = 0;
+
+//private:
+
+
+};
+
+
+#endif // ISPECULARSTRATEGY_H
