+++ Core/Multilayer/MatrixRTCoefficients_v2.h
@@ -26,7 +26,7 @@
 class BA_CORE_API_ MatrixRTCoefficients_v2 : public ILayerRTCoefficients
 {
 public:
-    friend class SpecularMagnetic_v2;
+    friend class SpecularMagneticStrategy;
 
     MatrixRTCoefficients_v2(double kz_sign, Eigen::Vector2cd eigenvalues, kvector_t b);
     MatrixRTCoefficients_v2(const MatrixRTCoefficients_v2& other);
